split install and uninstall
