+++ src/app/install-service/install-service.component.html
@@ -3,5 +3,4 @@
       {{ 'PAGES.INSTALL.TITLE' | translate }}
     </h1>
     <button mat-button (click)="install()" >install testscript</button>
-    <button mat-button (click)="uninstall()" >uninstall testscript</button>
   </div>
