+++ src/app/uninstall-service/uninstall-service.component.html
@@ -2,4 +2,5 @@
     <h1 class="title">
       {{ 'PAGES.UNINSTALL.TITLE' | translate }}
     </h1>
+    <button mat-button (click)="uninstall()" >uninstall testscript</button>
   </div>
